Refactor duplicated logic across codebase
- Extract solveInterceptQuadratic into src/utils/math.ts and reuse in steering.ts and aiming.ts
- Extract resolveMaterial into src/components/environment/starDisk/utils.ts and reuse in useStarDiskDebugCleanup.ts and useStarDiskFrameLoop.ts
- Remove deprecated physics wrappers (deferWrappers.ts, postWrappers.ts) and point imports to wrappers.js
