--- src/game/physics/deferWrappers.ts
@@ -1,18 +1,0 @@
-/**
- * Deferred physics mutation wrappers
- *
- * Re-exports from the consolidated wrappers module for backwards compatibility.
- * @deprecated Import directly from './wrappers.js' instead
- */
-
-export {
-  deferSetNextKinematicTranslation,
-  deferSetNextKinematicRotation,
-  deferSetLinvel,
-  deferSetAngvel,
-  deferSetMass,
-  deferSetLinearDamping,
-  deferSetAngularDamping,
-  deferSetColliderFriction,
-  deferSetColliderRestitution,
-} from './wrappers.js';
