+++ src/game/physics/index.ts
@@ -1,4 +1,3 @@
 export * from './types.js';
 export * from './mutationHelpers.js';
-export * from './deferWrappers.js';
-export * from './postWrappers.js';
+export * from './wrappers.js';
