--- src/game/physics/postWrappers.ts
@@ -1,18 +1,0 @@
-/**
- * Post-physics mutation wrappers
- *
- * Re-exports from the consolidated wrappers module for backwards compatibility.
- * @deprecated Import directly from './wrappers.js' instead
- */
-
-export {
-  postSetNextKinematicTranslation,
-  postSetNextKinematicRotation,
-  postSetLinvel,
-  postSetAngvel,
-  postSetMass,
-  postSetLinearDamping,
-  postSetAngularDamping,
-  postSetColliderFriction,
-  postSetColliderRestitution,
-} from './wrappers.js';
